+++ test/runtests.jl
@@ -14,6 +14,7 @@ using GLLVM
     include("test_packing.jl")
     include("test_fit.jl")
     include("test_fixed_effects.jl")
+    include("test_lv_predictor.jl")
     include("test_W_and_diag.jl")
     include("test_phy.jl")
     include("test_signed_sigma_phy.jl")
@@ -134,6 +135,7 @@ using GLLVM
     include("test_bridge_capabilities.jl")
     include("test_bridge_mixed.jl")
     include("test_bridge_x.jl")
+    include("test_bridge_lv_predictor.jl")
     include("test_bridge_missing_mask.jl")
     include("test_quality.jl")
 end